Split java inspectorprocessor stuff, create top level interface
